Made some changes to IMUSensor, created PTSensor, GPSSensor, and RTCSensor classes
I changed the IMUSensor so that all the getter methods were defined as inline in the header file. I also created the classes for the PT sensor, GPS, and RTC. I am having issues with getting the GPS to compile and I only started the RTC class, I'll look more into it later this week.